Moved fuzzer code into its own test module, and fixed to remove duplicate keys in fuzz batches
